+++ src/main/java/com/buildstore/common/validation/FieldMatchValidator.java
@@ -1,0 +1,30 @@
+package com.buildstore.common.validation;
+
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+import org.springframework.beans.BeanWrapperImpl;
+
+public class FieldMatchValidator implements ConstraintValidator<FieldMatch, Object> {
+
+    private String firstFieldName;
+    private String secondFieldName;
+
+    @Override
+    public void initialize(final FieldMatch constraintAnnotation) {
+        firstFieldName = constraintAnnotation.first();
+        secondFieldName = constraintAnnotation.second();
+    }
+
+    @Override
+    public boolean isValid(final Object value, final ConstraintValidatorContext context) {
+        try {
+            final Object firstObj = new BeanWrapperImpl(value).getPropertyValue(firstFieldName);
+            final Object secondObj = new BeanWrapperImpl(value).getPropertyValue(secondFieldName);
+
+            return firstObj == null && secondObj == null || firstObj != null && firstObj.equals(secondObj);
+        } catch (final Exception ignore) {
+            // ignore
+        }
+        return true;
+    }
+}